+++ src/Hmac/HmacCsrfToken.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Csrf\Stateless;
+namespace Yiisoft\Csrf\Hmac;
 
 use Yiisoft\Csrf\CsrfTokenInterface;
 use Yiisoft\Security\DataIsTamperedException;
@@ -13,13 +13,13 @@ use Yiisoft\Strings\StringHelper;
  * Stateless CSRF token does not require any storage. The token is a hash from session ID and a timestamp
  * (to prevent replay attacks). It is added to forms. When the form is submitted, we re-generate the token from
  * the current session ID and a timestamp from the original token. If two hashes match, we check that timestamp is
- * less than {@see StatelessCsrfToken::$lifetime}.
+ * less than {@see HmacCsrfToken::$lifetime}.
  *
  * The algorithm is also known as "HMAC Based Token".
  *
  * @see https://cheatsheetseries.owasp.org/cheatsheets/Cross-Site_Request_Forgery_Prevention_Cheat_Sheet.html#hmac-based-token-pattern
  */
-final class StatelessCsrfToken implements CsrfTokenInterface
+final class HmacCsrfToken implements CsrfTokenInterface
 {
     private CsrfTokenIdentityGeneratorInterface $identityGenerator;
     private Mac $mac;
