+++ tests/Mock/MockCsrfTokenIdentityGenerator.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Csrf\Tests\Mock;
 
-use Yiisoft\Csrf\Stateless\CsrfTokenIdentityGeneratorInterface;
+use Yiisoft\Csrf\Hmac\CsrfTokenIdentityGeneratorInterface;
 
 final class MockCsrfTokenIdentityGenerator implements CsrfTokenIdentityGeneratorInterface
 {
